fix warning

+++ zkSforce/zkAuthentication.m
@@ -198,7 +198,7 @@ static const int DEFAULT_MAX_SESSION_AGE = 25 * 60; // 25 minutes
 	self.sessionId = [lr sessionId];
 
 	// if we have a sessionSecondsValid in the UserInfo, use that to control when we re-authenticate, otherwise take the default.
-	int sessionAge = [[lr userInfo] sessionSecondsValid] > 0 ? [[lr userInfo] sessionSecondsValid] - 60 : DEFAULT_MAX_SESSION_AGE;
+	NSInteger sessionAge = [[lr userInfo] sessionSecondsValid] > 0 ? [[lr userInfo] sessionSecondsValid] - 60 : DEFAULT_MAX_SESSION_AGE;
 	self.sessionExpiresAt = [NSDate dateWithTimeIntervalSinceNow:sessionAge];
 	return lr;
 }
